requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
supabase==2.28.0
pydantic==2.12.5
python-dotenv==1.2.1
anthropic==0.83.0
httpx==0.27.2
apscheduler==3.11.2
aiosmtplib==5.1.0
jinja2==3.1.6
resend
fast-flights==2.2
